Beginning of the end (Huge improvements, technical challenge overcome)
FEAT: Began work on the hue slider's hue choosing mechanisms. It took slightly longer than I feel like it should've to figure this out, but I'm happy with its current implementation, and the ideas I have to improve it.
REFACTOR: Huge refactor on the color picker and how it is rendered. Firstly, the values are now almost not-hardcoded at all, being dynamic to the screen's gui size.
REFACTOR: Moved the whole color picker to its own class
REFACTOR: Removed the test color variable in the "aaaaaaaaa..." category
BUG: Color picker's x/xLimit seems to be flipped incorrectly
BUG: HUGE lag spike while choosing the value's hue(I'm unsure how I'm going to fix this)
TODO: Possible better implementation of the color picker as a whole to be clicked on not using the temporary workaround

This is more of a checkpoint for me than anything. I've spent multiple hours each night working on this, and losing this progress would be terrible.
